refactor(ui): Refactor all helpers
