+++ ui/src/helpers/vulnerability-statistics.ts
@@ -29,7 +29,7 @@ import {
   V4ScoreResult,
 } from 'ae-cvss-calculator/dist/types/src/CvssVector';
 
-import { Vulnerability } from '@/api/requests/types.gen';
+import { Vulnerability } from '@/hey-api';
 
 export type EpssData = {
   score: number;
